requirements.txt
asgiref
Django
djangorestframework
et
gunicorn
numpy
openpyxl
packaging
psycopg2
pytz
six
sqlparse
typing_extensions
tzdata
django-cors-headers
django-rest-framework
python-dateutil
python-dotenv
dj-database-url
